feat: moved cmd
