+++ db/db.go
@@ -1,7 +1,7 @@
 package db
 
 import (
-	"api/src/config"
+	"api/cmd/config"
 	"database/sql"
 	_ "github.com/go-sql-driver/mysql"
 	"log"
