--- src/validation/err.validation.go
@@ -1,13 +1,0 @@
-package validation
-
-import (
-	"api/src/response"
-	"net/http"
-)
-
-func Err(w http.ResponseWriter, status int, err error) {
-	if err != nil {
-		response.Err(w, status, err)
-		return
-	}
-}
